New background! and a small bug
** The small bug ** something's wrong with the camera in the menu scenes
(main menu and select screen). The camera just cuts some parts of the
background. I tried to fix it, but I suck...

Also, I waited the npc team to give me some pics of their new models,
but still nothing... So I guess we'll include that in the next sprint.